+++ example/package.json
@@ -10,6 +10,8 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^13.0.0",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "eslint-import-resolver-babel-module": "^5.3.2",
     "eslint-plugin-import": "^2.27.5",
     "expo": "^47.0.0",
@@ -17,7 +19,9 @@
     "expo-splash-screen": "^0.18.1",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "react-native-dotenv": "^3.3.1"
+    "react-native-dotenv": "^3.3.1",
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-screens": "^3.20.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
